+++ CHANGELOG.md
@@ -1301,3 +1301,4 @@
 - [2026-05-20 16:29] chore(auth): update dependencies and cleanup unused imports in auth
 - [2026-05-20 16:50] feat(test): add async test stream provider
 - [2026-05-20 16:53] refactor(perf): extract reusable atomic widgets from perf
+- [2026-05-20 17:52] style(perf): harmonize light and dark theme palette for perf